backend/requirements.txt
# Flask Backend Requirements
# Install with: pip install -r requirements.txt

Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Flask-Login==0.6.3
bcrypt==4.1.2
PyJWT==2.8.0
